+++ packages/autorest.typescript/src/transforms/mapperTransforms.ts
@@ -175,7 +175,9 @@ function buildMapper(
   const serializedName =
     getDiscriminatorValue(schema) ||
     options.serializedName ||
-    (schema as ObjectSchema).discriminator && !isUberParent(schema as ObjectSchema) && getLanguageMetadata(schema.language).name ||
+    ((schema as ObjectSchema).discriminator &&
+      !isUberParent(schema as ObjectSchema) &&
+      getLanguageMetadata(schema.language).name) ||
     // Fallback to name only for XML schemas since they need a name, otherwise don't
     (options.hasXmlMetadata && getLanguageMetadata(schema.language).name);
 
@@ -297,10 +299,10 @@ function buildAdditionalProperties(
   const additionalProperties = getAdditionalProperties(objectSchema);
   return additionalProperties
     ? {
-      type: {
-        name: MapperType.Object
+        type: {
+          name: MapperType.Object
+        }
       }
-    }
     : undefined;
 }
 
@@ -350,7 +352,9 @@ function transformObjectMapper(pipelineValue: PipelineValue) {
   );
 
   if (objectSchema.parents?.immediate[0]) {
-    uberParent = getMapperClassName(objectSchema.parents?.immediate[0] as ObjectSchema); 
+    uberParent = getMapperClassName(
+      objectSchema.parents?.immediate[0] as ObjectSchema
+    );
   }
   const mapper = buildMapper(
     schema,
@@ -593,7 +597,12 @@ function transformStringMapper(pipelineValue: PipelineValue) {
    */
   if (
     !isSchemaType(
-      [SchemaType.String, SchemaType.Credential, SchemaType.Uri],
+      [
+        SchemaType.String,
+        SchemaType.Credential,
+        SchemaType.Uri,
+        SchemaType.ArmId
+      ],
       schema
     )
   ) {
@@ -697,7 +706,9 @@ function processProperties(
     const propName = getLanguageMetadata(prop.language).name;
     const name = normalizeName(
       propName,
-      prop.language.default.isTopLevelParameter ? NameType.Parameter : NameType.Property,
+      prop.language.default.isTopLevelParameter
+        ? NameType.Parameter
+        : NameType.Property,
       true /** shouldGuard */
     );
     modelProperties[name] = getMapperOrRef(prop.schema, serializedName, {