+++ packages/cadl-rlc-test/test/customWrapper/cadl-output/src/parameters.ts
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { RequestParameters } from "@azure-rest/core-client";
+import { DeploymentCreationParameters } from "./models";
+
+export type GetDeploymentParameters = RequestParameters;
+
+export interface DeployProjectBodyParam {
+  /** Parameter of type 'DeploymentCreationParameters' in the body. */
+  body?: DeploymentCreationParameters;
+}
+
+export type DeployProjectParameters = DeployProjectBodyParam &
+  RequestParameters;