+++ packages/cadl-rlc-test/test/multiclient/cadl-output/src/core/coreClient.ts
@@ -6,10 +6,13 @@ import { CoreClient } from "./clientDefinitions";
 
 /**
  * Initialize a new instance of the class CoreClient class.
- *
+ * @param endpoint type: string
  */
-export function createClient(options: ClientOptions = {}): CoreClient {
-  const baseUrl = options.baseUrl ?? "undefined";
+export function createClient(
+  endpoint: string,
+  options: ClientOptions = {}
+): CoreClient {
+  const baseUrl = options.baseUrl ?? `${endpoint}`;
   options.apiVersion = options.apiVersion ?? "1.0.0";
 
   const userAgentInfo = `azsdk-js-multiclient-rest/1.0.0-beta.1`;