+++ packages/cadl-typescript/src/modelUtils.ts
@@ -10,8 +10,6 @@ import {
   getEffectiveModelType,
   getFormat,
   getFriendlyName,
-  getIntrinsicModelName,
-  getKnownValues,
   getMaxLength,
   getMaxValue,
   getMinLength,
@@ -21,7 +19,6 @@ import {
   getPropertyType,
   getSummary,
   getVisibility,
-  isIntrinsic,
   isNeverType,
   isNumericType,
   isSecret,
@@ -32,7 +29,9 @@ import {
   ModelProperty,
   Program,
   Type,
-  Union
+  Union,
+  isNullType,
+  Scalar
 } from "@cadl-lang/compiler";
 import { reportDiagnostic } from "./lib.js";
 import {
@@ -43,13 +42,15 @@ import {
   Schema,
   SchemaContext
 } from "@azure-tools/rlc-common";
+import { getResourceOperation } from "@cadl-lang/rest";
 import {
   getHeaderFieldName,
   getPathParamName,
   getQueryParamName,
-  isStatusCode
+  isStatusCode,
+  HttpOperation
 } from "@cadl-lang/rest/http";
-import { getPagedResult } from "@azure-tools/cadl-azure-core";
+import { getPagedResult, isFixed } from "@azure-tools/cadl-azure-core";
 
 export function getBinaryType(usage: SchemaContext[]) {
   return usage.includes(SchemaContext.Output)
@@ -63,7 +64,7 @@ export function getSchemaForType(
   usage?: SchemaContext[],
   needRef?: boolean
 ) {
-  const type = getEffectiveModelFromType(typeInput);
+  const type = getEffectiveModelFromType(program, typeInput);
   const builtinType = mapCadlTypeToTypeScript(program, type, usage);
   if (builtinType !== undefined) {
     // add in description elements for types derived from primitive types (SecureString, etc.)
@@ -85,22 +86,8 @@ export function getSchemaForType(
     return getSchemaForUnion(program, type, usage);
   } else if (type.kind === "Enum") {
     return getSchemaForEnum(program, type);
-  }
-  function getEffectiveModelFromType(type: Type) {
-    if (type.kind === "Model") {
-      const effective = getEffectiveModelType(program, type, isSchemaProperty);
-      if (effective.name) {
-        return effective;
-      }
-    }
-    function isSchemaProperty(property: ModelProperty) {
-      const headerInfo = getHeaderFieldName(program, property);
-      const queryInfo = getQueryParamName(program, property);
-      const pathInfo = getPathParamName(program, property);
-      const statusCodeInfo = isStatusCode(program, property);
-      return !(headerInfo || queryInfo || pathInfo || statusCodeInfo);
-    }
-    return type;
+  } else if (type.kind === "Scalar") {
+    return getSchemaForScalar(program, type);
   }
   if (isUnknownType(type)) {
     const returnType: any = { type: "unknown" };
@@ -120,6 +107,22 @@ export function getSchemaForType(
   });
   return undefined;
 }
+export function getEffectiveModelFromType(program: Program, type: Type): Type {
+  if (type.kind === "Model") {
+    const effective = getEffectiveModelType(program, type, isSchemaProperty);
+    if (effective.name) {
+      return effective;
+    }
+  }
+  function isSchemaProperty(property: ModelProperty) {
+    const headerInfo = getHeaderFieldName(program, property);
+    const queryInfo = getQueryParamName(program, property);
+    const pathInfo = getPathParamName(program, property);
+    const statusCodeInfo = isStatusCode(program, property);
+    return !(headerInfo || queryInfo || pathInfo || statusCodeInfo);
+  }
+  return type;
+}
 export function includeDerivedModel(model: Model): boolean {
   return (
     !isTemplateDeclaration(model) &&
@@ -129,11 +132,12 @@ export function includeDerivedModel(model: Model): boolean {
   );
 }
 
-function isNullType(program: Program, type: Type): boolean {
-  return (
-    isIntrinsic(program, type) &&
-    getIntrinsicModelName(program, type) === "null"
-  );
+function getSchemaForScalar(program: Program,  scalar: Scalar) {
+  let result = getSchemaForStdScalar(program, scalar);
+  if (!result && scalar.baseScalar) {
+    result = getSchemaForScalar(program, scalar.baseScalar);
+  }
+  return applyIntrinsicDecorators(program, scalar, result);
 }
 
 function getSchemaForUnion(
@@ -141,83 +145,32 @@ function getSchemaForUnion(
   union: Union,
   usage?: SchemaContext[]
 ) {
-  let type: string;
-  const nonNullOptions = union.options.filter((t) => !isNullType(program, t));
+  const nonNullOptions = union.options.filter((t) => !isNullType(t));
   const nullable = union.options.length != nonNullOptions.length;
   if (nonNullOptions.length === 0 || nonNullOptions[0] === undefined) {
     reportDiagnostic(program, { code: "union-null", target: union });
     return {};
   }
 
-  const kind = nonNullOptions[0].kind;
-  switch (kind) {
-    case "String":
-      type = "string";
-      break;
-    case "Number":
-      type = "number";
-      break;
-    case "Boolean":
-      type = "boolean";
-      break;
-    case "Model":
-      type = "model";
-      break;
-    default:
-      reportInvalidUnionForTypescript();
-      return {};
-  }
-
   const values = [];
-  if (type === "model" || type === "array") {
-    // Model unions can only ever be a model type with 'null'
-    if (nonNullOptions.length === 1) {
-      // Get the schema for the model type
-      const schema: any = getSchemaForType(program, nonNullOptions[0], usage);
-      if (schema) {
-        schema["x-nullable"] = nullable;
-      }
-
-      return schema;
-    } else {
-      reportDiagnostic(program, {
-        code: "union-unsupported",
-        messageId: "null",
-        target: union
-      });
-      return {};
-    }
-  }
 
   for (const option of nonNullOptions) {
-    if (option.kind != kind) {
-      reportInvalidUnionForTypescript();
-    }
-
     // We already know it's not a model type
-    values.push((option as any).value);
+    values.push(getSchemaForType(program, option, usage));
   }
 
-  const schema: any = { type };
+  const schema: any = {};
   if (values.length > 0) {
     schema.enum = values;
-    schema.type =
-      type === "string"
-        ? values.map((item) => `"${item}"`).join(" | ")
-        : values.join(" | ");
+    schema.type = values
+      .map((item) => `${getTypeName(item) ?? item}`)
+      .join(" | ");
   }
   if (nullable) {
-    schema["x-nullable"] = true;
+    schema["required"] = false;
   }
 
   return schema;
-
-  function reportInvalidUnionForTypescript() {
-    reportDiagnostic(program, {
-      code: "union-unsupported",
-      target: union
-    });
-  }
 }
 
 // An openapi "string" can be defined in several different ways in Cadl
@@ -551,8 +504,7 @@ function mapCadlTypeToTypeScript(
     case "Boolean":
       return { type: `${cadlType.value}` };
     case "Model":
-    case "ModelProperty":
-      return mapCadlIntrinsicModelToTypeScript(program, cadlType, usage);
+      return mapCadlStdTypeToTypeScript(program, cadlType, usage);
   }
   if (cadlType.kind === undefined) {
     if (typeof cadlType === "string") {
@@ -564,7 +516,7 @@ function mapCadlTypeToTypeScript(
 }
 function applyIntrinsicDecorators(
   program: Program,
-  cadlType: Model | ModelProperty,
+  cadlType: Scalar | ModelProperty,
   target: any
 ): any {
   const newTarget = { ...target };
@@ -621,17 +573,6 @@ function applyIntrinsicDecorators(
     newTarget["x-ms-secret"] = true;
   }
 
-  if (isString) {
-    const values = getKnownValues(program, cadlType);
-    if (values) {
-      const enumSchema = { ...newTarget, ...getSchemaForEnum(program, values) };
-      enumSchema.name = "string";
-      enumSchema.typeName = "string";
-
-      return enumSchema;
-    }
-  }
-
   return newTarget;
 }
 function getSchemaForEnum(program: Program, e: Enum) {
@@ -652,9 +593,12 @@ function getSchemaForEnum(program: Program, e: Enum) {
     schema.type =
       type === "string"
         ? values.map((item) => `"${item}"`).join("|")
-        : values.join("|");
+        : values.map((item) => `${item}`).join("|");
+    if (!isFixed(program, e)) {
+      schema.name = "string";
+      schema.typeName = "string";
+    }
   }
-
   return schema;
   function enumMemberType(member: EnumMember) {
     if (typeof member.value === "number") {
@@ -666,15 +610,15 @@ function getSchemaForEnum(program: Program, e: Enum) {
 /**
  * Map Cadl intrinsic models to open api definitions
  */
-function mapCadlIntrinsicModelToTypeScript(
+function mapCadlStdTypeToTypeScript(
   program: Program,
-  cadlType: Model | ModelProperty,
+  cadlType: Model,
   usage?: SchemaContext[]
 ): any | undefined {
   const indexer = (cadlType as Model).indexer;
   if (indexer !== undefined) {
     if (!isNeverType(indexer.key)) {
-      const name = getIntrinsicModelName(program, indexer.key);
+      const name = indexer.key.name;
       let schema: any = {};
       if (name === "string") {
         const valueType = getSchemaForType(
@@ -689,7 +633,7 @@ function mapCadlIntrinsicModelToTypeScript(
           description: getDoc(program, cadlType)
         };
         if (
-          !isIntrinsic(program, indexer.value) &&
+          !program.checker.isStdType(indexer.value) &&
           !isUnknownType(indexer.value!)
         ) {
           schema.typeName = `Record<string, ${valueType.name}>`;
@@ -713,7 +657,7 @@ function mapCadlIntrinsicModelToTypeScript(
           description: getDoc(program, cadlType)
         };
         if (
-          !isIntrinsic(program, indexer.value) &&
+          !program.checker.isStdType(indexer.value) &&
           !isUnknownType(indexer.value!) &&
           indexer.value?.kind &&
           schema.items.name
@@ -754,21 +698,24 @@ function mapCadlIntrinsicModelToTypeScript(
       return schema;
     }
   }
-  return getSchemaForIntrinsic(program, cadlType);
 }
 
-function getSchemaForIntrinsic(
+function getSchemaForStdScalar(
   program: Program,
-  cadlType: Model | ModelProperty
+  cadlType: Scalar
 ) {
-  if (!isIntrinsic(program, cadlType)) {
+  if (!program.checker.isStdType(cadlType)) {
     return undefined;
   }
-  const name = getIntrinsicModelName(program, cadlType);
+  const name = cadlType.name;
   const description = getSummary(program, cadlType);
   switch (name) {
     case "bytes":
       return { type: "string", format: "byte", description };
+    case "integer":
+      return applyIntrinsicDecorators(program, cadlType, {
+        type: "number"
+      });
     case "int8":
       return applyIntrinsicDecorators(program, cadlType, {
         type: "number",
@@ -904,4 +851,45 @@ export function getFormattedPropertyDoc(
     return `${propertyDoc}${sperator}${enhancedDocFromType}`;
   }
   return propertyDoc ?? enhancedDocFromType;
+}
+
+export function getBodyType(
+  program: Program,
+  route: HttpOperation
+): Type | undefined {
+  let bodyModel = route.parameters.bodyType;
+  if (bodyModel && bodyModel.kind === "Model" && route.operation) {
+    const resourceType = getResourceOperation(
+      program,
+      route.operation
+    )?.resourceType;
+    if (resourceType && route.responses && route.responses.length > 0) {
+      const resp = route.responses[0];
+      if (resp && resp.responses && resp.responses.length > 0) {
+        const responseBody = resp.responses[0]?.body;
+        if (responseBody) {
+          const bodyTypeInResponse = getEffectiveModelFromType(
+            program,
+            responseBody.type
+          );
+          // response body type is reosurce type, and request body type (if templated) contains resource type
+          if (
+            bodyTypeInResponse === resourceType &&
+            bodyModel.templateArguments &&
+            bodyModel.templateArguments.some((it) => {
+              return it.kind === "Model" || it.kind === "Union"
+                ? it === bodyTypeInResponse
+                : false;
+            })
+          ) {
+            bodyModel = resourceType;
+          }
+        }
+      }
+    }
+    if (resourceType && bodyModel.name === "") {
+      bodyModel = resourceType;
+    }
+  }
+  return bodyModel;
 }